IN-196 refactor: shortener imports
