+++ src/features/createPost/ui/cropping/ControllersPanel/AspectControllerButtons.tsx
@@ -8,7 +8,7 @@ import {
   SquarePhotoOutline,
 } from '@/assets/icons'
 import { Typography, cn, typographyVariants } from '@/common'
-import { addPostStore } from '@/features/createPost/model/addPostStore'
+import { addPostStore } from '@/features/createPost'
 import { observer } from 'mobx-react-lite'
 
 type ScaleSizeButtonType = {
