+++ src/features/createPost/ui/cropping/ControllersPanel/ScaleControllerPopover.tsx
@@ -3,9 +3,11 @@ import { useState } from 'react'
 
 import { Maximize, MaximizeOutline } from '@/assets/icons'
 import { Popover, PopoverContent, PopoverTrigger } from '@/common'
-import { addPostStore } from '@/features/createPost/model/addPostStore'
-import { Slider } from '@/features/createPost/ui/components'
-import { PhotoControllerButton } from '@/features/createPost/ui/components/photoControllerButton/PhotoControllerButton'
+import {
+  PhotoControllerButton,
+  Slider,
+  addPostStore,
+} from '@/features/createPost'
 import { observer } from 'mobx-react-lite'
 
 type Props = {
