+++ src/features/createPost/ui/filtering/FilterPhotoModal.tsx
@@ -9,9 +9,7 @@ import {
   cn,
   typographyVariants,
 } from '@/common'
-import { addPostStore } from '@/features/createPost/model/addPostStore'
-import { SwiperCover } from '@/features/createPost/ui/components/SwiperCover'
-import { Filters } from '@/features/createPost/ui/filtering/Filters'
+import { Filters, SwiperCover, addPostStore } from '@/features/createPost'
 import Image from 'next/image'
 import { SwiperSlide } from 'swiper/react'
